// web/src/lib/api/generated/models/tournamentResponse.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { AuditableResponse } from './auditableResponse';
import type { TournamentResponseFormat } from './tournamentResponseFormat';
import type { TournamentResponseStatus } from './tournamentResponseStatus';

export interface TournamentResponse {
  tournamentId: string;
  centerId: string;
  name: string;
  description?: string;
  format: TournamentResponseFormat;
  status: TournamentResponseStatus;
  maxPairs: number;
  startDate: string;
  endDate: string;
  audit?: AuditableResponse;
}